test for routine class funcs completed
